fw: Fix multiple synchronization issues with pulse timer
Reconfiguration is now only supported at the end of a pulse train and
the train timer is now updated appropriately. Reconfiguration in other
moments was mostly broken.
Stopping was refactored in order to eliminate the extra pulse added to
the pulse train in the previous implementation.
Critical regions were added to ensure correct operation when functions
are called concurrently.